Extract boot media dispatch helpers
