+++ crates/fstart-codegen/src/stage_gen/board_gen.rs
@@ -57,11 +57,13 @@ use fstart_types::{
 
 use super::tokens::hex_addr;
 
+mod boot_media;
 mod logger;
 mod model;
 mod phases;
 mod state;
 
+use boot_media::{anchor_bytes_stmt, match_boot_media};
 use logger::install_logger_body;
 use model::{device_provides, BoardCtx};
 use phases::{phase_init_body, PhaseSpec};
@@ -677,130 +679,6 @@ fn sig_verify_body(ctx: &BoardCtx<'_>) -> TokenStream {
 /// Emitted as a constant rather than a function because the token
 /// stream contains zero per-board variation.  Centralised so the
 /// safety comment lives in exactly one place.
-fn anchor_bytes_stmt() -> TokenStream {
-    quote! {
-        // SAFETY: FSTART_ANCHOR is emitted by
-        // `generate_anchor_static` in this same stage with proper
-        // alignment (`#[link_section = ".fstart.anchor"]` + `#[used]`)
-        // and is the size of `AnchorBlock`.  The FFS builder may
-        // patch its contents post-link, so downstream
-        // `FfsReader::read_anchor_volatile` reads it through
-        // `ptr::read_volatile`.
-        let _anchor_bytes: &[u8] = unsafe {
-            core::slice::from_raw_parts(
-                &FSTART_ANCHOR as *const fstart_types::ffs::AnchorBlock as *const u8,
-                core::mem::size_of::<fstart_types::ffs::AnchorBlock>(),
-            )
-        };
-    }
-}
-
-/// Emit a `match self._boot_media { ... }` that binds a local `_bm`
-/// for use by `bm_usage`.
-///
-/// Arms:
-///
-/// - `None` → `none_body` (caller decides: log + skip, or log + halt).
-/// - `Mmio { base, size }` → constructs `MemoryMapped` via
-///   `from_raw_addr`, then splices `bm_usage`.
-/// - `Block { device_id, offset, size }` → inner match over every
-///   enabled block device in the board, each arm building
-///   `BlockDeviceMedia` from `self.<field>` and splicing `bm_usage`.
-///   A wildcard arm logs and halts, labelled with `caller_tag` for
-///   the error message.
-///
-/// Caller passes `bm_usage` as a token stream that assumes `_bm` is
-/// in scope.  This keeps the control flow inside the emitted code
-/// (no closure crossing codegen / runtime) and lets a single helper
-/// service `sig_verify`, `fdt_prepare` Override, `payload_load`, and
-/// `stage_load` trampolines.
-fn match_boot_media(
-    ctx: &BoardCtx<'_>,
-    bm_usage: &TokenStream,
-    caller_tag: &str,
-    none_body: &TokenStream,
-) -> TokenStream {
-    let block_arms = block_device_arms(ctx, bm_usage);
-    let err_msg = format!("{caller_tag}: unknown block device id {{}}");
-    quote! {
-        match self._boot_media {
-            fstart_stage_runtime::BootMediaState::None => {
-                #none_body
-            }
-            fstart_stage_runtime::BootMediaState::Mmio { base, size } => {
-                fstart_log::info!("boot-media match: MMIO base={:#x} size={:#x}", base, size);
-                #[cfg(feature = "x86_64")]
-                fstart_platform::enable_boot_media_rom_cache();
-                // SAFETY: `base..base + size` is the board's
-                // memory-mapped flash window; the board author's
-                // RON declared this range is readable.
-                let _bm = unsafe {
-                    fstart_services::boot_media::MemoryMapped::from_raw_addr(
-                        base,
-                        size as usize,
-                    )
-                };
-                fstart_log::info!("boot-media match: media constructed");
-                #bm_usage
-            }
-            fstart_stage_runtime::BootMediaState::Block {
-                device_id,
-                offset,
-                size,
-            } => {
-                match device_id {
-                    #block_arms
-                    _ => {
-                        // Executor invariant: `boot_media_static(Some(id), ..)`
-                        // is only called with a device id that provides
-                        // `BlockDevice`.  Reaching this arm means the plan
-                        // carries an id we didn't emit a field for — a
-                        // codegen bug.  `unreachable_unchecked` would be
-                        // sound here (executor guarantees) but `halt` is
-                        // strictly safer.
-                        fstart_log::error!(#err_msg, device_id);
-                        fstart_platform::halt();
-                    }
-                }
-            }
-        }
-    }
-}
-
-/// Emit one match arm per enabled block device in the board.
-///
-/// Each arm constructs a
-/// [`BlockDeviceMedia`](fstart_services::boot_media::BlockDeviceMedia)
-/// from the corresponding `self.<name>` field (which the executor
-/// has guaranteed is `Some` via a prior `init_device` call) and
-/// splices `bm_usage` — a token stream that uses the `_bm` local.
-///
-/// Returns an empty token stream when the board has no block
-/// devices — the `match` arm set then contains only `_` and the
-/// `None` / `Mmio` arms.
-fn block_device_arms(ctx: &BoardCtx<'_>, bm_usage: &TokenStream) -> TokenStream {
-    let arms = enabled_indices(ctx.devices, ctx.instances, ctx.excluded)
-        .filter(|idx| device_provides(ctx, *idx, Service::BlockDevice))
-        .map(|idx| {
-            let dev = &ctx.devices[idx];
-            let field = format_ident!("{}", dev.name.as_str());
-            let id_lit = proc_macro2::Literal::u8_unsuffixed(idx as u8);
-            quote! {
-                #id_lit => {
-                    let _bm = fstart_services::boot_media::BlockDeviceMedia::new(
-                        self.#field
-                            .as_ref()
-                            .unwrap_or_else(|| fstart_platform::halt()),
-                        offset,
-                        size as usize,
-                    );
-                    #bm_usage
-                }
-            }
-        });
-    quote! { #(#arms)* }
-}
-
 /// Emit the body of `Board::fdt_prepare`.
 ///
 /// Mirrors the old `capabilities::generate_fdt_prepare` logic, but
